Extract EntryLocaleTransformer
